re-organized the file system and added a main.py function
