+++ requirements.txt
@@ -1,0 +1,7 @@
+scapy>=2.5.0
+python-nmap>=0.7.1
+pandas>=1.5.0
+numpy>=1.23.0
+pytest>=7.3.1
+black>=22.3.0
+pylint>=2.15.0
